desktop/requirements.txt
# 三角洲战术终端 - 桌面客户端依赖

# GUI框架
PyQt6>=6.4.0

# 屏幕捕获
mss>=9.0.0
Pillow>=10.0.0

# Windows API
pywin32>=306

# 进程检测
psutil>=5.9.0

# OCR引擎 (选择其一)
# paddleocr>=2.7.0  # 推荐，中文识别效果好
# paddlepaddle>=2.5.0  # PaddleOCR依赖
easyocr>=1.7.0  # 备选，安装简单
# pytesseract>=0.3.10  # 需要额外安装Tesseract

# 图像处理
numpy>=1.24.0
opencv-python>=4.8.0

# 数据处理
pandas>=2.0.0

# HTTP请求（用于数据同步）
requests>=2.31.0

# 热键支持
keyboard>=0.13.5

# 系统托盘
# pystray>=0.19.0  # 可选
